+++ deepomatic/dmake/loader/v0_1/__init__.py
@@ -1,1 +1,1 @@
-from deepobuild import DMakeFile
+from .dmake_file import DMakeFile
