+++ install.sh
@@ -130,7 +130,7 @@ echo "export DMAKE_PATH=${DMAKE_PATH}" >> ${CONFIG_FILE}
 echo "export DMAKE_CONFIG_DIR=${DMAKE_CONFIG_DIR}" >> ${CONFIG_FILE}
 echo "export DMAKE_SSH_KEY=${DMAKE_SSH_KEY}" >> ${CONFIG_FILE}
 echo "export PYTHONPATH=\$PYTHONPATH:${DMAKE_PATH}" >> ${CONFIG_FILE}
-echo "export PATH=\$PATH:${DMAKE_PATH}/deepomatic/dmake/:${DMAKE_PATH}/deepomatic/dmake/utils" >> ${CONFIG_FILE}
+echo "export PATH=\$PATH:${DMAKE_PATH}:${DMAKE_PATH}/deepomatic/dmake/utils" >> ${CONFIG_FILE}
 
 LINE="source ${CONFIG_FILE}"
 if [ -z "`which dmake`" ]; then
@@ -149,4 +149,7 @@ if [ -z "`which dmake`" ]; then
 else
     echo "Patched config to version ${DMAKE_VERSION}"
 fi
+
+# Install libs
+pip install -r $DMAKE_PATH/requirements.txt
